Split web review controls tests
