+++ apps/web/src/screens/review/tests/ReviewScreen.controls.editor.test.tsx
@@ -1,0 +1,118 @@
+// @vitest-environment jsdom
+import { describe, expect, it, vi } from "vitest";
+import type { Card } from "../../../types";
+import {
+  clickElementAsync,
+  createCard,
+  loadReviewQueueSnapshotMock,
+  setTextFieldValueAsync,
+  setupReviewScreenTest,
+} from "../testSupport/ReviewScreenTestSupport";
+
+const {
+  dispatchDocumentKeydown,
+  getContainer,
+  getState,
+  renderReviewScreen,
+  revealAnswer,
+} = setupReviewScreenTest();
+
+describe("ReviewScreen editor controls", () => {
+  it("saves card edits from the review editor", async () => {
+    const state = getState();
+    const card = createCard({
+      cardId: "card-edit",
+      frontText: "Before",
+      backText: "Existing back",
+      tags: ["grammar"],
+    });
+    state.cards = [card];
+    state.reviewQueue = [card];
+    state.reviewTimeline = [card];
+
+    await renderReviewScreen();
+
+    const editButton = getContainer().querySelector(".review-pane-edit-btn");
+    if (!(editButton instanceof HTMLButtonElement)) {
+      throw new Error("Review edit button was not found");
+    }
+
+    await clickElementAsync(editButton);
+
+    const frontTextField = document.getElementById("review-card-editor-front-text");
+    if (!(frontTextField instanceof HTMLTextAreaElement)) {
+      throw new Error("Review editor front field was not found");
+    }
+
+    await setTextFieldValueAsync(frontTextField, "After");
+
+    const saveButton = [...document.querySelectorAll(".review-editor-modal .primary-btn")][0];
+    if (!(saveButton instanceof HTMLButtonElement)) {
+      throw new Error("Review editor save button was not found");
+    }
+
+    await clickElementAsync(saveButton);
+
+    expect(state.appData.updateCardItem).toHaveBeenCalledWith("card-edit", {
+      frontText: "After",
+      backText: "Existing back",
+      tags: ["grammar"],
+    });
+  });
+
+  it("deletes the edited card after confirmation", async () => {
+    const state = getState();
+    const card = createCard({
+      cardId: "card-delete",
+    });
+    state.cards = [card];
+    state.reviewQueue = [card];
+    state.reviewTimeline = [card];
+    const confirmMock = vi.spyOn(window, "confirm").mockReturnValue(true);
+
+    await renderReviewScreen();
+
+    const editButton = getContainer().querySelector(".review-pane-edit-btn");
+    if (!(editButton instanceof HTMLButtonElement)) {
+      throw new Error("Review edit button was not found");
+    }
+
+    await clickElementAsync(editButton);
+
+    const deleteButton = document.querySelector(".review-editor-delete-btn");
+    if (!(deleteButton instanceof HTMLButtonElement)) {
+      throw new Error("Review editor delete button was not found");
+    }
+
+    await clickElementAsync(deleteButton);
+
+    expect(confirmMock).toHaveBeenCalledWith("Delete this card?");
+    expect(state.appData.deleteCardItem).toHaveBeenCalledWith("card-delete");
+
+    confirmMock.mockRestore();
+  });
+
+  it("keeps rating shortcuts disabled until the answer is visible", async () => {
+    const state = getState();
+    const card = createCard({
+      cardId: "card-hidden-answer",
+      frontText: "Question",
+      backText: "Answer",
+    });
+    state.cards = [card];
+    state.reviewQueue = [card];
+    state.reviewTimeline = [card];
+    state.appData.submitReviewItem.mockImplementation(async (): Promise<Card> => card);
+    loadReviewQueueSnapshotMock.mockClear();
+
+    await renderReviewScreen();
+    await dispatchDocumentKeydown("1");
+
+    expect(state.appData.submitReviewItem).not.toHaveBeenCalled();
+
+    await revealAnswer();
+    await dispatchDocumentKeydown("1");
+
+    expect(state.appData.submitReviewItem).toHaveBeenCalledWith("card-hidden-answer", 0);
+  });
+});
